+++ Assets/UniGLTF/Editor/MeshUtility/BoneMeshEraserGUI.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 5cdbec3de528c61448aeb739ffe1f79e
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
