+++ Assets/UniGLTF/Editor/MeshUtility/MeshProcessingMessages.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 575d573f28f0e4a48a2a6509e69b7f6f
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
